requirements.txt
discord.py == 2.3.2
requests == 2.31.0
Pillow == 9.5.0
google-generativeai == 0.3.2
requests == requests
regex == 2.2.1
